requirements.txt
azure-core==1.34.0 ; python_version >= "3.11"
azure-functions==1.23.0 ; python_version >= "3.11"
azure-storage-blob==12.25.1 ; python_version >= "3.11"
certifi==2025.6.15 ; python_version >= "3.11"
cffi==1.17.1 ; python_version >= "3.11" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.2 ; python_version >= "3.11"
cryptography==45.0.4 ; python_version >= "3.11"
idna==3.10 ; python_version >= "3.11"
isodate==0.7.2 ; python_version >= "3.11"
markupsafe==3.0.2 ; python_version >= "3.11"
pycparser==2.22 ; python_version >= "3.11" and platform_python_implementation != "PyPy"
requests==2.32.4 ; python_version >= "3.11"
six==1.17.0 ; python_version >= "3.11"
typing-extensions==4.14.0 ; python_version >= "3.11"
urllib3==2.5.0 ; python_version >= "3.11"
werkzeug==3.1.3 ; python_version >= "3.11"
